QOL Changes
I didn't quite understand what I was seeing in the progress tracking so I made it more verbose and informative. Also, sometimes, this extension, during loading with some other extensions, throws an error. I haven't quite found out which other extensions specifically but I do know that nothing is actually wrong. The extension still functions in the UI. So, I told it to just keep going so it loads silently.